Tests/LibWeb/Text/expected/HTML/formAction-attribute.txt

button.formAction initial value: http://www.example.com/
Final segment of button.formAction after setting to the empty string: formAction-attribute.html
Final segment of button.formAction after setting to "../test.html": test.html
input.formAction initial value: http://www.example.com/
Final segment of input.formAction after setting to the empty string: formAction-attribute.html
Final segment of input.formAction after setting to "../test.html": test.html
